fixing all tapi kontak outlet belum dicantumin masih dalam pengujian

+++ database/migrations/2023_03_24_174137_create_outlets_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name_outlet');
             $table->string('slug');
             $table->string('kantor')->nullable();
-            $table->integer('kontak')->nullable();
+            $table->bigInteger('kontak')->nullable();
             $table->string('jadwal')->nullable();
             $table->string('foto_outlet')->nullable();
             $table->timestamps();
